+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ### Added
 
+- Added service `Services\Catalog\ProductPropertyEnum` with support methods,
+  see [catalog.productPropertyEnum.* methods](https://apidocs.bitrix24.com/api-reference/catalog/product-property-enum/index.html) ([#549](https://github.com/bitrix24/b24phpsdk/issues/549)):
+    - `add` creates a new list-type property value
+    - `update` updates an existing list-type property value
+    - `get` gets a list-type property value by identifier
+    - `list` gets the list of list-type property values by filter
+    - `delete` deletes a list-type property value by identifier
+    - `getFields` returns the description of list-type property value fields
 - Added service `Services\Landing\Site\Service\Site` with support methods,
   see [landing.site.* methods](https://github.com/bitrix24/b24phpsdk/issues/267):
     - `add` adds a site
@@ -155,6 +163,13 @@
   `start`, `pause`, `defer`, `complete`, etc.) for users migrating to the v3 SDK.
   All classes under `Bitrix24\SDK\Legacy\` are marked `@deprecated` and will be removed
   once v3 reaches feature parity with v1.
+
+### Fixed
+
+- Fixed batch operations for `Services\Catalog\Product` using the wrong-case `ID` key instead of
+  the lowercase `id` key expected by `catalog.product.list` and `catalog.product.delete`: added
+  `Services\Catalog\Product\Batch` overriding `determineKeyId()` and `deleteEntityItems()`,
+  registered in `CatalogServiceBuilder::product()` ([#549](https://github.com/bitrix24/b24phpsdk/issues/549))
 
 ## 3.0.0 - 2026.01.01